back/requirements.txt
fastapi
uvicorn
pydantic
passlib
bcrypt
python-dotenv
python-jose[cryptography]
python-multipart
passlib[bcrypt]
